nx503a: msm: mdss: apply BW fudge factor to all controllers when video mode present
cherry-pick from 5e19c8365084747deabda631da24fd566b6ef1b1

If there is a video mode controller enabled and other
controllers non-video mode, current display driver apply
the bw fudge factor only to the controller that it is
video mode.
This change will apply the fudge factor to all the
controllers enabled if any of those it is video mode to
fix some of the underruns observed in the video interfaces.