nx503a: defconfig: Disable CONFIG_PFT as it is unsupported
Disable the drivers/platform/msm/pft.c driver as it is unsupported by
Qualcomm, and opens us up to a wide range of potential attack surfaces
that has not been audited by anyone.

Qualcomm recommends that it be disabled, as it hooks into SELinux in
some "interesting" ways, and the userspace portion of the code is not
even part of the image.

Bug: 28588434
Qualcomm-case-number: 02440206

This reverts commit fa4f7c1fd235914909b2764cc79bca816f4b87ef.
http://review.cyanogenmod.org/#/c/153688/